requirements.txt
# Core FL Framework
flwr[simulation]==1.7.0
flower==1.7.0

# Deep Learning
torch==2.1.0
torchvision==0.16.0
torchaudio==0.16.0

# Data & Utilities
numpy==1.24.3
scipy==1.11.0
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.8.0
seaborn==0.13.0
Pillow==10.0.0

# Optimization & Privacy
PyYAML==6.0
tqdm==4.66.1
tensorboard==2.14.0

# Gradient Inversion Specific
opencv-python==4.8.0.76
scikit-image==0.21.0

# Development
jupyter==1.0.0
ipython==8.14.0
black==23.9.1
flake8==6.1.0